Put WalletHomeViewNoCredentials component into separate file
